requirements.txt
Django==4.1.3
django-rest-framework==0.1.0
requests==2.31.0